requirements.txt
Flask==2.3.2
openai==0.27.2
requests==2.31.0
ffmpeg-python==0.2.0
google-cloud-texttospeech==2.14.1
gunicorn==21.2.0
python-dotenv==1.0.0